fix: load module defaults and variables if available
